refactor: remove UserState files/hooks from lending
